refactor(ai): split treasury_planner into part-of concern files (#3288)
Issue #3288 maintainability slice: the 1428-physical-line
treasury_planner.dart exceeds the 1000-line code-review guideline noted
in the issue investigation. Extract it into single-concern `part of`
fragments; the main library drops to 498 lines and no fragment exceeds
~280 lines.

New part files (same library scope — no import/visibility changes):
- treasury_need_analysis.dart (surplus/need maps, forecasting,
  carry-forward, speculative-bid pass)
- treasury_market_pricing.dart (production-cost gate, bid-priority tiers,
  liquidity commodity, supplier offer-tier alignment)
- treasury_lock_recovery.dart (lock-recovery seller predicate, designated/
  liquidity buyer rotation, liquidity bid)
- treasury_regiment_bootstrap.dart (regiment build-input feedstock/
  improvement-input bootstrap helpers)
- treasury_bid_emission.dart (cargo/treasury-budget-clamped bid prioritization)

Behaviour-preserving: code moved verbatim into part fragments that share
the parent library's imports and private scope, so symbols, visibility,
and determinism are unchanged. The parent retains the public entry point
(runTreasuryPlanner), cargo resolver, and regiment-rebuild orchestrator
plus the five new part directives.

Remaining #3288 perf slices (snapshot-threaded province counts, OrdersBuilder,
hot-path log guards) are out of scope for this slice.

Refs #3288
