+++ packages/colonizethis_ai/lib/src/planning/treasury_planner.dart
@@ -20,6 +20,15 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'army_conquest_prep.dart' show regimentCountForPlayer;
 import 'recipe_scoring.dart' show kShortageThreshold;
 
+// Treasury planner concern fragments (Refs #3288 file-split). Each `part of`
+// fragment shares this library's imports and private scope, so the move is
+// behaviour-preserving — symbols, visibility, and helper sharing are unchanged.
+part 'treasury_need_analysis.dart';
+part 'treasury_market_pricing.dart';
+part 'treasury_lock_recovery.dart';
+part 'treasury_regiment_bootstrap.dart';
+part 'treasury_bid_emission.dart';
+
 /// Bid priority tiers (1 = highest). Refs #2994 F4.
 const int kTreasuryBidPriorityEssentialInput = 1;
 const int kTreasuryBidPriorityLuxury = 2;
@@ -486,937 +495,4 @@ void _applyLockRecoverySellerRegimentRebuildBids({
       need: need,
     );
   }
-}
-
-Stockpile _projectStockpileAfterProduction({
-  required Stockpile stockpile,
-  required List<AssignedRecipe> productionAssignments,
-}) {
-  var projected = stockpile;
-  for (final assignment in productionAssignments) {
-    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
-    if (recipe == null || assignment.assignedLabour <= 0) continue;
-    final runs = assignment.assignedLabour ~/ recipe.labourPerOutput;
-    if (runs <= 0) continue;
-    for (final entry in recipe.inputQuantities.entries) {
-      projected = projected.applyDelta(entry.key, -entry.value * runs);
-    }
-    projected = projected.applyDelta(
-      recipe.outputCommodityId,
-      recipe.outputQuantity * runs,
-    );
-  }
-  return projected;
-}
-
-Set<CommodityId> _trackedCommodityIds({
-  required Stockpile stockpile,
-  required Stockpile projected,
-  required Map<CommodityId, int> inputNeeds,
-  required List<AssignedRecipe> productionAssignments,
-}) {
-  final ids = <CommodityId>{...inputNeeds.keys};
-  for (final entry in stockpile.quantities.entries) {
-    if (entry.value > 0) ids.add(entry.key);
-  }
-  for (final entry in projected.quantities.entries) {
-    if (entry.value > 0) ids.add(entry.key);
-  }
-  for (final assignment in productionAssignments) {
-    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
-    if (recipe == null) continue;
-    ids.add(recipe.outputCommodityId);
-    ids.addAll(recipe.inputQuantities.keys);
-  }
-  for (final commodity in CommodityCatalog.all) {
-    if (commodity.category == CommodityCategory.food) {
-      ids.add(commodity.id);
-    }
-  }
-  return ids;
-}
-
-int _consumptionForecast({
-  required CommodityId commodityId,
-  required Commodity commodity,
-  required Map<CommodityId, int> inputNeeds,
-}) {
-  if (inputNeeds.containsKey(commodityId)) {
-    return inputNeeds[commodityId]!.clamp(1, kShortageThreshold);
-  }
-  if (commodity.category == CommodityCategory.food) {
-    return kShortageThreshold;
-  }
-  return (kShortageThreshold ~/ 2).clamp(1, kShortageThreshold);
-}
-
-Map<CommodityId, int> _inputNeedsFromAssignments(
-  List<AssignedRecipe> productionAssignments,
-) {
-  final needs = <CommodityId, int>{};
-  for (final assignment in productionAssignments) {
-    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
-    if (recipe == null || assignment.assignedLabour <= 0) continue;
-    final runs = assignment.assignedLabour ~/ recipe.labourPerOutput;
-    if (runs <= 0) continue;
-    for (final entry in recipe.inputQuantities.entries) {
-      needs[entry.key] = (needs[entry.key] ?? 0) + entry.value * runs;
-    }
-  }
-  return needs;
-}
-
-bool _marketPriceBelowProductionCost(
-  CommodityId commodityId,
-  Map<CommodityId, int> marketPrices,
-) {
-  final marketPrice = marketPrices[commodityId];
-  if (marketPrice == null) return true;
-  var bestCost = double.infinity;
-  for (final recipe in ProductionRecipesCatalog.producing(commodityId)) {
-    var inputCost = 0.0;
-    for (final entry in recipe.inputQuantities.entries) {
-      final inputPrice = marketPrices[entry.key] ?? 0;
-      inputCost += inputPrice * entry.value;
-    }
-    final perUnit = inputCost / recipe.outputQuantity;
-    if (perUnit < bestCost) bestCost = perUnit;
-  }
-  if (bestCost == double.infinity) return true;
-  return marketPrice < bestCost;
-}
-
-int _bidPriorityForCommodity(CommodityId commodityId) {
-  final commodity = CommodityCatalog.byId[commodityId];
-  if (commodity == null) return kTreasuryBidPriorityRawMaterial;
-  return switch (commodity.category) {
-    CommodityCategory.manufactured ||
-    CommodityCategory.advanced =>
-      kTreasuryBidPriorityEssentialInput,
-    CommodityCategory.food => kTreasuryBidPriorityFood,
-    CommodityCategory.rawMaterial => kTreasuryBidPriorityRawMaterial,
-    CommodityCategory.luxury => kTreasuryBidPriorityLuxury,
-    CommodityCategory.riches => kTreasuryBidPriorityRawMaterial,
-  };
-}
-
-/// Returns prior-turn offer-side fill rate (`0.0`–`1.0`) for [commodityId] from
-/// [state.lastTurnActivity]. Returns `1.0` when no activity exists or
-/// `totalOfferQuantity <= 0`, matching the "no prior data → assume fully
-/// fillable" convention used elsewhere in F2/F3. Refs #2994 F8.
-double _priorTurnOfferFillRate(WorldMarketState state, CommodityId commodityId) {
-  final activity = state.lastTurnActivity[commodityId];
-  if (activity == null) return 1.0;
-  final total = activity.totalOfferQuantity;
-  if (total <= 0) return 1.0;
-  final fillFraction = activity.filledQuantity / total;
-  if (fillFraction.isNaN || !fillFraction.isFinite) return 1.0;
-  if (fillFraction < 0.0) return 0.0;
-  if (fillFraction > 1.0) return 1.0;
-  return fillFraction;
-}
-
-/// Carry-forward residual quantity per commodity for [playerId] on the given
-/// [side], summed across all carry-forward entries. Returns an empty map when
-/// the player has no carry-forward state. Refs #2994 F8.
-Map<CommodityId, int> _carryForwardQuantitiesByCommodity({
-  required WorldMarketState state,
-  required String playerId,
-  required TradeOrderType side,
-}) {
-  final source = switch (side) {
-    TradeOrderType.offer => state.carryForwardOffersByFactionId[playerId],
-    TradeOrderType.bid => state.carryForwardBidsByFactionId[playerId],
-  };
-  if (source == null || source.isEmpty) {
-    return const <CommodityId, int>{};
-  }
-  final result = <CommodityId, int>{};
-  for (final order in source) {
-    if (order.quantity <= 0) continue;
-    result[order.commodityId] =
-        (result[order.commodityId] ?? 0) + order.quantity;
-  }
-  return result;
-}
-
-/// True when some faction other than [excludePlayerId] holds a standing
-/// (carry-forward) offer for [commodityId] in the world market — i.e. real
-/// supply the buyer can bid against. Refs #2847 H8-supply castIron source:
-/// once an affluent supplier's over-produced `castIron` surplus stands in the
-/// market, a locked seller bids `castIron` **directly** instead of routing to
-/// domestic production from feedstock it cannot extract. Pure read of
-/// [WorldMarketState.carryForwardOffersByFactionId]; deterministic.
-bool _marketHasStandingOfferSupplyFromOthers({
-  required WorldMarketState state,
-  required CommodityId commodityId,
-  required String excludePlayerId,
-}) {
-  for (final entry in state.carryForwardOffersByFactionId.entries) {
-    if (entry.key == excludePlayerId) continue;
-    for (final order in entry.value) {
-      if (order.commodityId == commodityId && order.quantity > 0) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-/// Discounts forecasted treasury inflow from this turn's offers by the
-/// prior-turn offer-side fill rate per commodity, rounded to an integer
-/// treasury unit. A first-ever market turn (no prior activity) returns the
-/// nominal sum (full-fill credit) per the F2/F3 default. Refs #2994 F8.
-int _expectedOfferInflow({
-  required Map<CommodityId, int> available,
-  required Map<CommodityId, int> marketPrices,
-  required WorldMarketState state,
-}) {
-  if (available.isEmpty) return 0;
-  var inflow = 0.0;
-  for (final entry in available.entries) {
-    final commodityId = entry.key;
-    final quantity = entry.value;
-    if (quantity <= 0) continue;
-    final price = marketPrices[commodityId] ?? 0;
-    if (price <= 0) continue;
-    final fillRate = _priorTurnOfferFillRate(state, commodityId);
-    inflow += quantity * price * fillRate;
-  }
-  if (!inflow.isFinite) return 0;
-  return inflow.round();
-}
-
-/// Speculative-bid pass for affluent GPs (Refs #2924 F10). Mutates [need] in
-/// place with synthetic stockpile-target deficits for non-riches commodities
-/// the F1–F5 path did not already speak for. Adds **at most one** entry per
-/// invocation so bids are concentrated on the commodity most likely to clear
-/// into a real deal (treasury only redistributes when matching offers exist).
-/// Selection order:
-/// 1. Commodities with prior-turn `MarketActivity.totalOfferQuantity > 0`
-///    (descending offer volume, then alphabetical) — proven liquidity.
-/// 2. Otherwise food commodities (deterministic alphabetical) — minor/tribe
-///    auto-offers reliably surface food on the next world-market phase.
-/// 3. Otherwise the alphabetical first non-riches commodity that meets the
-///    target gap (deterministic fallback for an empty market — pure
-///    determinism for tests that do not seed `lastTurnActivity`).
-/// Skips:
-/// - riches commodities (excluded from world-market trade),
-/// - commodities already in [need] (F1–F5 deficit path owns them),
-/// - commodities already in [available] (mutual-exclusion preserved),
-/// - commodities whose projected stockpile already meets
-///   [kSpeculativeBidStockpileTarget] (no positive target),
-/// - commodities whose remaining target is already covered by a carry-forward
-///   bid residual ([carryForwardBids]).
-void _addSpeculativeBidNeeds({
-  required Map<CommodityId, int> need,
-  required Map<CommodityId, int> available,
-  required Stockpile projected,
-  required Map<CommodityId, int> carryForwardBids,
-  required WorldMarketState state,
-}) {
-  bool eligible(CommodityId id) {
-    if (richesCommodityIds.contains(id)) return false;
-    if (need.containsKey(id)) return false;
-    if (available.containsKey(id)) return false;
-    final projectedQty = projected.quantityOf(id);
-    final carryQty = carryForwardBids[id] ?? 0;
-    return kSpeculativeBidStockpileTarget - projectedQty - carryQty > 0;
-  }
-
-  int gapFor(CommodityId id) {
-    final projectedQty = projected.quantityOf(id);
-    final carryQty = carryForwardBids[id] ?? 0;
-    return kSpeculativeBidStockpileTarget - projectedQty - carryQty;
-  }
-
-  int offerVolumeFor(CommodityId id) =>
-      state.lastTurnActivity[id]?.totalOfferQuantity ?? 0;
-
-  final eligibleIds = CommodityCatalog.all
-      .map((c) => c.id)
-      .where(eligible)
-      .toList(growable: false);
-  if (eligibleIds.isEmpty) return;
-
-  CommodityId pick;
-  final liquid = eligibleIds.where((id) => offerVolumeFor(id) > 0).toList()
-    ..sort((a, b) {
-      final volCmp = offerVolumeFor(b).compareTo(offerVolumeFor(a));
-      if (volCmp != 0) return volCmp;
-      return a.compareTo(b);
-    });
-  if (liquid.isNotEmpty) {
-    pick = liquid.first;
-  } else {
-    final foods = eligibleIds
-        .where(
-          (id) => CommodityCatalog.byId[id]?.category == CommodityCategory.food,
-        )
-        .toList()
-      ..sort();
-    if (foods.isNotEmpty) {
-      pick = foods.first;
-    } else {
-      final sortedEligible = [...eligibleIds]..sort();
-      pick = sortedEligible.first;
-    }
-  }
-  need[pick] = gapFor(pick);
-}
-
-/// Food commodity with the highest prior-turn offer volume on the world
-/// market; used as the lock-recovery bid target when every GP is selling
-/// surplus food under urgent offers. Refs #2924 F11.
-CommodityId _lockRecoveryLiquidityCommodity(WorldMarketState state) {
-  CommodityId? bestId;
-  var bestVolume = 0;
-  for (final entry in state.lastTurnActivity.entries) {
-    final commodityId = entry.key;
-    final commodity = CommodityCatalog.byId[commodityId];
-    if (commodity == null || commodity.category != CommodityCategory.food) {
-      continue;
-    }
-    final volume = entry.value.totalOfferQuantity;
-    if (volume > bestVolume) {
-      bestVolume = volume;
-      bestId = commodityId;
-      continue;
-    }
-    if (volume == bestVolume &&
-        bestId != null &&
-        commodityId.compareTo(bestId) < 0) {
-      bestId = commodityId;
-    }
-  }
-  if (bestId != null) return bestId;
-  final foods = CommodityCatalog.all
-      .where((c) => c.category == CommodityCategory.food)
-      .map((c) => c.id)
-      .toList(growable: false)
-    ..sort();
-  return foods.isNotEmpty ? foods.first : 'grain';
-}
-
-int _treasuryForPlayer(Game game, String playerId) =>
-    game.playerById(playerId)?.treasury ?? 0;
-
-int _newWorldProvinceCountOwnedBy(Game game, String playerId) {
-  var count = 0;
-  for (final province in game.worldState.newWorld.provinces) {
-    if (province.ownerId == playerId) count++;
-  }
-  return count;
-}
-
-/// Below-quota GPs with zero NW provinces and at least one OW province
-/// are Path F lock-recovery **sellers** — they must accumulate seller
-/// credits toward the regiment threshold, not rotate as the affluent
-/// designated buyer or speculate (Refs #2924 Path F gp6 regression).
-bool _isBelowQuotaZeroNwLockRecoverySeller({
-  required Game game,
-  required String playerId,
-}) {
-  final ow = oldWorldProvinceCountOwnedBy(game, playerId);
-  if (ow <= 0) return false;
-  if (!isBelowObserverConquestQuota(ow)) return false;
-  if (_newWorldProvinceCountOwnedBy(game, playerId) != 0) return false;
-  // Mid-below-quota EXPAND band (seed-42 gp3–gp6); excludes minimal
-  // single-province test fixtures that are not Path F lock-recovery sellers.
-  return ow >= 2;
-}
-
-/// Feedstock commodity ids consumed by production recipes whose output is a
-/// currently-missing cheapest-regiment (`peasant_levies`) build input.
-///
-/// Refs #2847 § H8-supply. The cheapest regiment needs `fabric`, which is
-/// produced from `wool` (`fabricFromWool`) or `cotton` (`fabricFromCotton`).
-/// When the projected stockpile is short of a build input, this returns the
-/// input commodities of every recipe that outputs that build input so a
-/// lock-recovery seller can retain the feedstock instead of selling it as
-/// surplus. Pure function of the projected [Stockpile] and the static
-/// `RegimentEconomyCatalog` / `ProductionRecipesCatalog`; returns the empty
-/// set once every build input is on hand (the reservation self-clears).
-Set<CommodityId> _regimentBuildInputFeedstockIds(Stockpile projected) {
-  final missingInputs = <CommodityId>{
-    for (final entry
-        in RegimentEconomyCatalog.peasantLevies.buildInputs.entries)
-      if (projected.quantityOf(entry.key) < entry.value) entry.key,
-  };
-  if (missingInputs.isEmpty) return const <CommodityId>{};
-  final feedstock = <CommodityId>{};
-  for (final buildInputId in missingInputs) {
-    for (final recipe in ProductionRecipesCatalog.producing(buildInputId)) {
-      feedstock.addAll(recipe.inputQuantities.keys);
-    }
-  }
-  return feedstock;
-}
-
-/// Build-input feedstock ids ordered so Old World lock-recovery sellers extract
-/// wool before cotton (seed-42), then alphabetically for determinism.
-List<CommodityId> _sortedRegimentBuildInputFeedstockIds(Stockpile projected) {
-  return _regimentBuildInputFeedstockIds(projected).toList()
-    ..sort((a, b) {
-      if (a == CommodityCatalog.wool.id) return -1;
-      if (b == CommodityCatalog.wool.id) return 1;
-      return a.compareTo(b);
-    });
-}
-
-/// Adds the feedstock bid for the first viable [feedstockCandidates] entry so a
-/// lock-recovery seller can domestically produce a missing regiment build input.
-/// Returns true when a feedstock bid was queued (still accumulating feedstock).
-bool _addRegimentBuildInputFeedstockBootstrapNeed({
-  required List<CommodityId> feedstockCandidates,
-  required Stockpile projected,
-  required Map<CommodityId, int> carryForwardBids,
-  required Map<CommodityId, int> need,
-}) {
-  for (final feedstockId in feedstockCandidates) {
-    final qtyNeeded =
-        _feedstockQuantityForOneMissingBuildInputRun(feedstockId, projected);
-    if (qtyNeeded <= 0) continue;
-    final held =
-        projected.quantityOf(feedstockId) + (carryForwardBids[feedstockId] ?? 0);
-    if (held >= qtyNeeded) return false;
-    need[feedstockId] = qtyNeeded - held;
-    return true;
-  }
-  return false;
-}
-
-/// Adds bids for the level-0 `build_improvement` inputs (lumber + cast iron) a
-/// lock-recovery seller must hold to extract its owned fabric feedstock tile,
-/// but does not, when the regiment build-input feedstock-extraction gate is
-/// active and a feedstock tile is owned (Refs #2847 § H8-extraction).
-///
-/// Returns true when at least one improvement-input bid was queued (a deficit
-/// remains). The caller suppresses the downstream feedstock / fabric bootstrap
-/// bids on that turn so the single `bidTypeCap` slot targets the prerequisite
-/// supply. Self-clearing: the
-/// [regimentBuildInputFeedstockImprovementInputCost] contract returns the empty
-/// map once the GP owns a regiment, holds the build input, or has improved the
-/// feedstock tile.
-bool _addRegimentFeedstockImprovementInputNeed({
-  required Game game,
-  required String playerId,
-  required Stockpile projected,
-  required Map<CommodityId, int> carryForwardBids,
-  required Map<CommodityId, int> need,
-}) {
-  final cost = regimentBuildInputFeedstockImprovementInputCost(game, playerId);
-  if (cost.isEmpty) return false;
-  // Deterministic iteration over the cost map (insertion order is fixed by the
-  // contract, but sorting guards against future reordering).
-  final inputIds = cost.keys.toList()..sort();
-
-  // Pass 1 — directly-buyable improvement-inputs (e.g. `lumber`, which suppliers
-  // release as surplus). These are acquired first so the single `bidTypeCap`
-  // slot targets an essential input the world market can actually supply (the
-  // suggester admits bids in alphabetical, cap-bounded order, so a raw-material
-  // feedstock bid would otherwise crowd out the essential `lumber` bid).
-  var queued = false;
-  for (final inputId in inputIds) {
-    // Refs #2847 H8-supply castIron source: a domestic-production improvement
-    // input (e.g. `castIron`) is normally produced from feedstock (Pass 2)
-    // because the world market structurally lacks supply. Once an affluent
-    // supplier over-produces and offers it (a standing carry-forward offer
-    // exists), bid it **directly** here so the deal can cross — breaking the
-    // extraction deadlock without waiting on feedstock the seller cannot mine.
-    if (kDomesticProductionImprovementInputIds.contains(inputId) &&
-        !_marketHasStandingOfferSupplyFromOthers(
-          state: game.worldMarketState,
-          commodityId: inputId,
-          excludePlayerId: playerId,
-        )) {
-      continue;
-    }
-    final held =
-        projected.quantityOf(inputId) + (carryForwardBids[inputId] ?? 0);
-    final missing = cost[inputId]! - held;
-    if (missing > 0) {
-      need[inputId] = missing;
-      queued = true;
-    }
-  }
-  if (queued) return true;
-
-  // Pass 2 — domestically-produced improvement-inputs (Refs #2847 H8-extraction
-  // castIron residual). `castIron` has no world-market supply on seed 42 (it is
-  // consumed by Old World military builds, so no Great Power offers a surplus),
-  // so the seller bids `castIron`'s production feedstock (`timber` + `iron`) and
-  // the economy planner produces it (SPEC/ai/treasury-planner.md § Lock-recovery
-  // seller improvement-input domestic production). A still-missing
-  // domestic-production input keeps the fabric/feedstock bootstrap suppressed
-  // even when its feedstock is already on hand (production is pending).
-  for (final inputId in inputIds) {
-    if (!kDomesticProductionImprovementInputIds.contains(inputId)) continue;
-    final held =
-        projected.quantityOf(inputId) + (carryForwardBids[inputId] ?? 0);
-    if (cost[inputId]! - held <= 0) continue;
-    _addImprovementInputProductionFeedstockNeed(
-      improvementInputId: inputId,
-      projected: projected,
-      carryForwardBids: carryForwardBids,
-      need: need,
-    );
-    queued = true;
-  }
-  return queued;
-}
-
-/// Bids the production feedstock for one run of [improvementInputId] when a
-/// lock-recovery seller is short of the feedstock to produce it domestically.
-/// Adds nothing when the feedstock for one full run is already on hand (the
-/// economy planner will run the recipe next; the caller still suppresses the
-/// direct improvement-input bid because the world market structurally lacks
-/// supply for it). Deterministic: selects the lowest-`id` recipe producing the
-/// input. Refs #2847 § H8-extraction castIron residual.
-void _addImprovementInputProductionFeedstockNeed({
-  required CommodityId improvementInputId,
-  required Stockpile projected,
-  required Map<CommodityId, int> carryForwardBids,
-  required Map<CommodityId, int> need,
-}) {
-  final recipe = _lowestIdRecipeProducing(improvementInputId);
-  if (recipe == null) return;
-  for (final entry in recipe.inputQuantities.entries) {
-    final held =
-        projected.quantityOf(entry.key) + (carryForwardBids[entry.key] ?? 0);
-    final missing = entry.value - held;
-    if (missing > 0) {
-      need[entry.key] = (need[entry.key] ?? 0) + missing;
-    }
-  }
-}
-
-/// The production recipe with the lowest `id` whose output is [commodityId], or
-/// `null` when no recipe produces it. Deterministic over the static
-/// `ProductionRecipesCatalog`.
-ProductionRecipe? _lowestIdRecipeProducing(CommodityId commodityId) {
-  ProductionRecipe? best;
-  for (final recipe in ProductionRecipesCatalog.producing(commodityId)) {
-    if (best == null || recipe.id.compareTo(best.id) < 0) {
-      best = recipe;
-    }
-  }
-  return best;
-}
-
-/// Adds direct bids for any missing `peasant_levies` build input when no
-/// feedstock bootstrap bid is pending.
-void _addRegimentBuildInputDirectNeed({
-  required Stockpile projected,
-  required Map<CommodityId, int> carryForwardBids,
-  required Map<CommodityId, int> need,
-}) {
-  for (final input in RegimentEconomyCatalog.peasantLevies.buildInputs.entries) {
-    final held =
-        projected.quantityOf(input.key) + (carryForwardBids[input.key] ?? 0);
-    if (held < input.value) {
-      need[input.key] = input.value - held;
-    }
-  }
-}
-
-/// Per-run feedstock input quantity required to produce one unit of a missing
-/// `peasant_levies` build input via a production recipe consuming [feedstockId].
-int _feedstockQuantityForOneMissingBuildInputRun(
-  CommodityId feedstockId,
-  Stockpile projected,
-) {
-  var needed = 0;
-  for (final entry in RegimentEconomyCatalog.peasantLevies.buildInputs.entries) {
-    if (projected.quantityOf(entry.key) >= entry.value) continue;
-    for (final recipe in ProductionRecipesCatalog.producing(entry.key)) {
-      final perRun = recipe.inputQuantities[feedstockId];
-      if (perRun != null && perRun > needed) {
-        needed = perRun;
-      }
-    }
-  }
-  return needed;
-}
-
-/// Level-0 `build_improvement` improvement-input commodities a lock-recovery
-/// seller must produce domestically because the world market structurally lacks
-/// supply for them on seed 42. `castIron` is consumed by Old World military
-/// builds, so no Great Power offers a surplus for the seller's bid to cross
-/// (SPEC/ai/treasury-planner.md § Lock-recovery seller improvement-input
-/// domestic production; Refs #2847 § H8-extraction castIron residual). For these
-/// inputs the seller bids the production feedstock (e.g. `timber` + `iron` for
-/// `castIron`) and the economy planner produces the input
-/// (economy-planner.md § Regiment build-input production priority).
-const Set<CommodityId> kDomesticProductionImprovementInputIds = {'castIron'};
-
-/// Commodity ids affluent GPs release when any lock-recovery seller needs the
-/// H8 bootstrap path (Refs #2847 H8-supply market).
-const Set<CommodityId> _regimentBuildInputSupplyCommodityIds = {
-  'wool',
-  'cotton',
-  'fabric',
-  // Refs #2847 H8-extraction: the level-0 build_improvement inputs the locked
-  // seller bids for to unblock domestic feedstock extraction.
-  'lumber',
-  'castIron',
-};
-
-/// Re-tags a lock-recovery supplier's build-input supply offers so each lands
-/// in the **same** integer priority tier the locked buyer bids that commodity
-/// at (`_bidPriorityForCommodity`), enabling the per-commodity offer/bid cross
-/// the DealMatcher otherwise blocks across mismatched tiers
-/// (SPEC/program/world-market-resolution.md § Step C; Refs #2847 H8-supply
-/// market order matching). Only offers whose commodity is in
-/// [_regimentBuildInputSupplyCommodityIds] are retuned; all other offers (for
-/// example the urgent liquidity-food offer) keep their computed priority. The
-/// original list is returned unchanged when no tier actually moves so equal
-/// inputs keep their identity (determinism).
-List<TradeOrder> _alignBuildInputSupplyOfferTiers(List<TradeOrder> offers) {
-  if (offers.isEmpty) return offers;
-  var changed = false;
-  final result = <TradeOrder>[];
-  for (final offer in offers) {
-    if (!_regimentBuildInputSupplyCommodityIds.contains(offer.commodityId)) {
-      result.add(offer);
-      continue;
-    }
-    final alignedPriority = _bidPriorityForCommodity(offer.commodityId);
-    if (alignedPriority == offer.priority) {
-      result.add(offer);
-      continue;
-    }
-    result.add(offer.copyWith(priority: alignedPriority));
-    changed = true;
-  }
-  return changed ? result : offers;
-}
-
-/// True when any below-quota zero-NW lock-recovery seller still needs the
-/// H8 regiment build-input bootstrap path (Refs #2847 H8-supply market).
-bool _anyLockRecoverySellerNeedsRegimentBuildInput(Game game) {
-  final threshold = cheapestRegimentBuildTreasuryCost();
-  for (final player in game.players) {
-    if (!_isBelowQuotaZeroNwLockRecoverySeller(
-      game: game,
-      playerId: player.id,
-    )) {
-      continue;
-    }
-    if (player.treasury < threshold) continue;
-    if (regimentCountForPlayer(game, player.id) > 0) continue;
-    for (final entry
-        in RegimentEconomyCatalog.peasantLevies.buildInputs.entries) {
-      if (player.stockpile.quantityOf(entry.key) < entry.value) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-/// Public accessor for the below-quota zero-NW lock-recovery seller predicate
-/// (Refs #2847 H8-supply castIron source). The economy planner uses it to keep
-/// the supplier `castIron` over-production off for a GP that is itself a locked
-/// seller (its own self-path boost already covers it).
-bool isBelowQuotaZeroNwLockRecoverySeller(Game game, String playerId) =>
-    _isBelowQuotaZeroNwLockRecoverySeller(game: game, playerId: playerId);
-
-/// True when any below-quota zero-NW lock-recovery seller still lacks a
-/// domestically-produced level-0 `build_improvement` input (a
-/// [kDomesticProductionImprovementInputIds] commodity such as `castIron`) the
-/// world market structurally cannot supply on seed 42 (Refs #2847 H8-supply
-/// castIron source). When true, an affluent supplier over-produces that input
-/// for release (`economy_planner.dart`) and this planner releases the resulting
-/// surplus + aligns its offer tier so the locked seller's bid can cross.
-/// Pure function of `(game)` and the static catalogs; returns `false` once no
-/// locked seller still needs the improvement input (self-clearing).
-bool anyLockRecoverySellerNeedsCastIronImprovementInput(Game game) {
-  for (final player in game.players) {
-    if (!_isBelowQuotaZeroNwLockRecoverySeller(
-      game: game,
-      playerId: player.id,
-    )) {
-      continue;
-    }
-    final cost = regimentBuildInputFeedstockImprovementInputCost(
-      game,
-      player.id,
-    );
-    if (cost.isEmpty) continue;
-    for (final entry in cost.entries) {
-      if (!kDomesticProductionImprovementInputIds.contains(entry.key)) {
-        continue;
-      }
-      if (player.stockpile.quantityOf(entry.key) < entry.value) {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-/// Sorted Great Power ids for deterministic per-turn buyer rotation.
-List<String> _sortedGreatPowerIds(Game game) {
-  final ids = <String>[
-    for (final player in game.players) player.id,
-  ]..sort();
-  return ids;
-}
-
-/// True iff at least one Great Power has `player.treasury <
-/// cheapestRegimentBuildTreasuryCost()`. The lock-recovery liquidity bid is
-/// only useful when at least one broke GP needs buy-side demand for its
-/// urgent offers. Refs #2924 F12.
-bool _anyBrokeGreatPower(Game game) {
-  final threshold = cheapestRegimentBuildTreasuryCost();
-  for (final player in game.players) {
-    if (player.treasury < threshold) return true;
-  }
-  return false;
-}
-
-bool _isAffluentDesignatedLockRecoveryBuyer({
-  required Game game,
-  required String playerId,
-}) {
-  if (!_anyBrokeGreatPower(game)) return false;
-  final gpIds = _sortedGreatPowerIds(game);
-  final affluent = <String>[
-    for (final id in gpIds)
-      if (_treasuryForPlayer(game, id) >= treasuryAffluenceThreshold() &&
-          !_isBelowQuotaZeroNwLockRecoverySeller(game: game, playerId: id))
-        id,
-  ];
-  if (affluent.isEmpty) return false;
-  final designated = lockRecoveryDesignatedBuyerId(game);
-  return designated.isNotEmpty && playerId == designated;
-}
-
-/// Whether [playerId] should emit the urgent lock-recovery liquidity-food bid
-/// this turn. Refs #2924 F11/F12/F13/F15.
-bool isLockRecoveryLiquidityBuyer({
-  required Game game,
-  required String playerId,
-  required int treasuryBudgetForBids,
-  required int treasuryForecast,
-}) {
-  if (!_anyBrokeGreatPower(game)) return false;
-  final liquidity = _lockRecoveryLiquidityCommodity(game.worldMarketState);
-  final pricePerUnit = game.worldMarketState.prices[liquidity] ?? 0;
-  if (pricePerUnit <= 0 || treasuryBudgetForBids < pricePerUnit) {
-    return false;
-  }
-  final threshold = cheapestRegimentBuildTreasuryCost();
-  final rawTreasury = _treasuryForPlayer(game, playerId);
-  // F13: optimistic offer-inflow forecast keeps a broke GP on offers-only.
-  if (rawTreasury < threshold && treasuryForecast >= threshold) {
-    return false;
-  }
-  final gpIds = _sortedGreatPowerIds(game);
-  final affluent = <String>[
-    for (final id in gpIds)
-      if (_treasuryForPlayer(game, id) >= treasuryAffluenceThreshold() &&
-          !_isBelowQuotaZeroNwLockRecoverySeller(game: game, playerId: id))
-        id,
-  ];
-  if (affluent.isNotEmpty) {
-    final designated = lockRecoveryDesignatedBuyerId(game);
-    return designated.isNotEmpty && playerId == designated;
-  }
-  // F15: when no GP is affluent, logic-phase minor auto-bids (`world_market_phase`
-  // / `computeLockRecoveryMinorAutoBids`) fund liquidity-food purchases. GP buyers
-  // would spend scarce treasury on grain instead of accumulating seller credits.
-  return false;
-}
-
-/// Preferred liquidity buyers when no GP is affluent (6-GP observer order).
-/// gp1/gp2 exit EXPAND earlier on seed 42 than gp3–gp6; keeping buys on these
-/// factions prevents stuck EXPAND sellers from spending their own treasury.
-const List<String> kLockRecoveryPreferredBuyerIds = ['gp1', 'gp2'];
-
-/// Buyer when no GP meets [treasuryAffluenceThreshold]: rotate among
-/// [kLockRecoveryPreferredBuyerIds] present in the game, else the two
-/// richest-by-treasury GPs.
-String lockRecoveryFallbackBuyerId(Game game) {
-  final gpIds = _sortedGreatPowerIds(game);
-  if (gpIds.isEmpty) return '';
-  final preferred = <String>[
-    for (final id in kLockRecoveryPreferredBuyerIds)
-      if (gpIds.contains(id)) id,
-  ];
-  final buyerPool = preferred.length >= 2
-      ? preferred
-      : _twoRichestGreatPowerIdsByTreasury(game);
-  if (buyerPool.isEmpty) return '';
-  if (buyerPool.length == 1) return buyerPool.first;
-  final turn = game.worldState.turnState.turnNumber;
-  return buyerPool[turn % buyerPool.length];
-}
-
-List<String> _twoRichestGreatPowerIdsByTreasury(Game game) {
-  final gpIds = _sortedGreatPowerIds(game);
-  if (gpIds.isEmpty) return const [];
-  final ranked = [...gpIds]
-    ..sort((a, b) {
-      final tA = _treasuryForPlayer(game, a);
-      final tB = _treasuryForPlayer(game, b);
-      if (tA != tB) return tB.compareTo(tA);
-      return a.compareTo(b);
-    });
-  return ranked.take(2).toList();
-}
-
-/// One GP per turn acts as the market buyer for the lock-recovery food
-/// commodity so other GPs' urgent offers can clear. Refs #2924 F11.
-///
-/// Rotates only among GPs at or above [treasuryAffluenceThreshold] so a
-/// broke designated buyer does not waste the single `bidTypeCap` slot on
-/// grain bids its treasury cannot fund while other GPs' urgent offers
-/// starve (seed-42 gp4/gp6 `totalDealsAsSeller == 0` diagnostic). When no
-/// GP meets the affluence band, [isLockRecoveryLiquidityBuyer] admits every
-/// GP whose per-turn bid budget can fund at least one liquidity-food unit
-/// (F15) instead of using this rotation.
-///
-/// Returns the empty string when no Great Power is broke — every GP is
-/// already at or above the regiment threshold and the F1–F5 / F10 paths
-/// handle the steady state without a synthetic grain bid. Refs #2924 F12.
-String lockRecoveryDesignatedBuyerId(Game game) {
-  if (!_anyBrokeGreatPower(game)) return '';
-  final gpIds = _sortedGreatPowerIds(game);
-  if (gpIds.isEmpty) return '';
-  final threshold = treasuryAffluenceThreshold();
-  final affluent = <String>[
-    for (final id in gpIds)
-      if (_treasuryForPlayer(game, id) >= threshold &&
-          !_isBelowQuotaZeroNwLockRecoverySeller(game: game, playerId: id))
-        id,
-  ];
-  if (affluent.isEmpty) return '';
-  final turn = game.worldState.turnState.turnNumber;
-  return affluent[turn % affluent.length];
-}
-
-/// Designated buyer bids [commodityId] and does not offer it this turn.
-///
-/// Bid quantity is the smaller of the F11 stockpile-target ceiling and
-/// `max(0, treasuryBudgetForBids / pricePerUnit)` so the buyer never commits
-/// more treasury than it currently holds (Refs #2924 F12 — treasury-capped)
-/// **and** never overcommits against pending costs or carry-forward bid
-/// notional already accounted for in [treasuryBudgetForBids] (Refs #3122).
-void _applyLockRecoveryLiquidityBid({
-  required String playerId,
-  required Game game,
-  required Map<CommodityId, int> need,
-  required Map<CommodityId, int> available,
-  required Map<CommodityId, int> carryForwardBids,
-  required int treasuryBudgetForBids,
-  required int treasuryForecast,
-  required bool addSyntheticBid,
-}) {
-  final commodityId = _lockRecoveryLiquidityCommodity(game.worldMarketState);
-  available.remove(commodityId);
-  if (!addSyntheticBid) return;
-  final pricePerUnit = game.worldMarketState.prices[commodityId] ?? 0;
-  if (pricePerUnit <= 0) return;
-  final budget =
-      treasuryBudgetForBids < 0 ? 0 : treasuryBudgetForBids;
-  final affordableQty = budget ~/ pricePerUnit;
-  // Refs #2924 F14: lock-recovery liquidity bids use the full per-turn
-  // treasury budget (after pending costs and carry-forward notional), not
-  // the F10 stockpile-target ceiling of 8 units. On seed 42 the designated
-  // buyer's treasury is far below the affluent band; capping at 8 kept per-
-  // deal seller credits too small to approach the regiment threshold.
-  final liquidityQty = affordableQty;
-  if (liquidityQty <= 0) return;
-  final existing = need[commodityId] ?? 0;
-  if (liquidityQty > existing) {
-    need[commodityId] = liquidityQty;
-  }
-}
-
-List<TradeOrder> _prioritizedBids({
-  required List<TradeOrder> rawBids,
-  required Map<CommodityId, int> need,
-  required int bidTypeCap,
-  required int tradeCargoCapacity,
-  required int offerPriority,
-  required bool alignBidPriorityWithUrgentOffers,
-  required int treasuryBudgetForBids,
-  required WorldMarketState worldMarketState,
-  required ResourceRules resourceRules,
-  int? forceBidPriority,
-  CommodityId? preferCommodityId,
-}) {
-  if (rawBids.isEmpty || bidTypeCap <= 0 || tradeCargoCapacity <= 0) {
-    return const <TradeOrder>[];
-  }
-  final byCommodity = <CommodityId, TradeOrder>{
-    for (final bid in rawBids) bid.commodityId: bid,
-  };
-  final orderedIds = need.keys.toList(growable: false)
-    ..sort((a, b) {
-      if (preferCommodityId != null) {
-        if (a == preferCommodityId) return -1;
-        if (b == preferCommodityId) return 1;
-      }
-      final priorityCmp =
-          _bidPriorityForCommodity(a).compareTo(_bidPriorityForCommodity(b));
-      if (priorityCmp != 0) return priorityCmp;
-      return a.compareTo(b);
-    });
-
-  final result = <TradeOrder>[];
-  var remainingCargo = tradeCargoCapacity;
-  var remainingTreasuryBudget =
-      treasuryBudgetForBids < 0 ? 0 : treasuryBudgetForBids;
-  var admitted = 0;
-  for (final commodityId in orderedIds) {
-    if (admitted >= bidTypeCap) break;
-    if (remainingCargo <= 0) break;
-    final bid = byCommodity[commodityId];
-    if (bid == null) continue;
-    final cargoClampedQty = bid.quantity < remainingCargo
-        ? bid.quantity
-        : remainingCargo;
-    if (cargoClampedQty <= 0) continue;
-    // Refs #3122: clamp every bid to the running treasury budget so the
-    // matcher (#3115) does not have to truncate to zero/near-zero fills.
-    // When the commodity has no effective price (manufactured commodity
-    // before in-game price discovery seeds a price), fall back to the
-    // cargo-clamped quantity; the matcher applies its own per-tier
-    // accounting if such a bid ever clears.
-    final pricePerUnit = effectiveMarketPriceForCommodityId(
-      commodityId: commodityId,
-      worldMarket: worldMarketState,
-      resourceRules: resourceRules,
-    );
-    int cappedQty;
-    if (pricePerUnit == null) {
-      cappedQty = cargoClampedQty;
-    } else if (pricePerUnit <= 0) {
-      cappedQty = cargoClampedQty;
-    } else {
-      final maxAffordable = remainingTreasuryBudget ~/ pricePerUnit;
-      cappedQty = cargoClampedQty < maxAffordable
-          ? cargoClampedQty
-          : maxAffordable;
-    }
-    if (cappedQty <= 0) continue;
-    result.add(
-      bid.copyWith(
-        quantity: cappedQty,
-        priority: forceBidPriority ??
-            (alignBidPriorityWithUrgentOffers
-                ? offerPriority
-                : _bidPriorityForCommodity(commodityId)),
-      ),
-    );
-    remainingCargo -= cappedQty;
-    if (pricePerUnit != null && pricePerUnit > 0) {
-      remainingTreasuryBudget -= cappedQty * pricePerUnit;
-    }
-    admitted += 1;
-  }
-  return result;
 }
